Add a simple seed-set expansion routine.
The routine is in the wrong place (json utils), adds a total weight into
STINGER that likely is best maintained separately, etc.  The algorithn
is O(|S|^2 + volume of one more step out) where |S| is the size of the
output.  Not pretty.

Amusingly, this runs concurrently with changes to the STINGER data
structure.  Because we have only insertions, this is ok.  Deletions,
well, that's kinda an open problem.  The later routine that picks up the
set and lists the edges may have to cope.

Maintaining the total weight separately can work entirely on the batch
for insertions, but not removals:

  - Given a batch of insertions / positive increments:
    - Just add the weights.
  - Given a batch of removals / decrements:
    - Need to know the edge weights being removed.

STING batch preprocessing should (eventually) annotate the edge removals
and weight decrements with existing edge information that summarizes all
the edges the requested types to be removed/decremented.